requirements.txt
websocket-client==1.6.4
kafka-python==2.0.2
python-dotenv==1.0.0
requests==2.31.0 